this is really a base 36 encoder and not a base 32 encoder given the 26 letters in the alphabet and the 10 numbers being used in this encoding
