+++ Core/Core.csproj
@@ -108,7 +108,7 @@
     <Compile Include="Services\CSharpValidator.cs" />
     <Compile Include="Data\IPostRepository.cs" />
     <Compile Include="Services\MissingSemicolonRewriter.cs" />
-    <Compile Include="Utilities\Base32Encoder.cs" />
+    <Compile Include="Utilities\Base36Encoder.cs" />
     <Compile Include="Utilities\TaskAsyncHelper.cs" />
   </ItemGroup>
   <ItemGroup>
